Variable throughput experiment script (memcached)
Update memcached scripts to support background worker contending with
server.

+++ exp/variable_throughput/run_memcached_exp.sh
@@ -2,6 +2,12 @@
 # This script is executed on Mutilate master machine!
 
 # GLOBAL VARIABLES
+exp_name=("Error" "ETC" "Fixed Value No BG" "Fixed Value BG")
+ETC=1
+FIXED_VALUE_NO_BG=2
+FIXED_VALUE_BG=3
+EXPERIMENT=$ETC
+
 server_ip=node0
 # server_port=11211
 server_cores=32
@@ -38,6 +44,7 @@ function check_env {
 
 function run_memcached_server {
   pkill memcached
+  sleep 1
   a=`memcached -V`
   if [ $? -ne 0 ]
   then
@@ -52,10 +59,33 @@ function run_memcached_server {
     sudo systemctl disable memcached
   fi
 
+  if [ $EXPERIMENT -eq $FIXED_VALUE_BG ]
+  then
+    # Check if sysbench is installed
+    a=`sysbench --version`
+    if [ $? -ne 0 ]
+    then
+      sudo apt install -y sysbench
+      if [ $? -ne 0 ]
+      then
+        echo "Failed to install sysbench on the server"
+        exit 1
+      fi
+    fi
+  fi
+
   memory=50000 # MB
   connections=2048
   server_cmd="taskset -c 1-$server_cores memcached -l $server_ip -m $memory -c $connections -t $server_cores -d"
-  exec sh -c "$server_cmd"
+  exec sh -c "$server_cmd" &
+  echo $server_cmd
+  if [ $EXPERIMENT -eq $FIXED_VALUE_BG ]
+  then
+    echo "Running Background Program (sysbenc)" 
+    # --rate=10000 
+    taskset -c 1-$server_cores sysbench --threads=$server_cores \
+      --time=$((mutilate_duration + mutilate_warmup + 10)) --forced-shutdown=2 cpu run
+  fi
 }
 
 function install_mutilate {
@@ -79,8 +109,12 @@ function install_mutilate {
 }
 
 function run_agent {
-  rm $FINISH_FILE
   pkill mutilate
+  sleep 1
+  if [ -e $FINISH_FILE ]
+  then 
+    rm $FINISH_FILE
+  fi
   cd $HOME
   install_mutilate
   cd mutilate
@@ -91,7 +125,7 @@ function run_agent {
 }
 
 function run_master {
-  pkill mutilate
+  # pkill mutilate
   cd $HOME
   install_mutilate
   cd mutilate
@@ -102,7 +136,19 @@ function run_master {
   done
   echo "Agents: $agents"
   # key_value_data="--keysize=fb_key --valuesize=fb_value --iadist=fb_ia --update=0.033"
-  key_value_data="--keysize=fb_key --valuesize=pareto:0,2144,0.348238 --iadist=fb_ia --update=0.033"
+  case $EXPERIMENT in
+    $ETC)
+      key_value_data="--keysize=fb_key --valuesize=pareto:0,2144,0.348238 --iadist=fb_ia --update=0.033"
+      ;;
+    $FIXED_VALUE_NO_BG)
+      # TODO: It is duplicate of FIXED_VALUE_BG
+      key_value_data="--keysize=20 --valuesize=4800"
+      ;;
+    $FIXED_VALUE_BG)
+      # TODO: It is duplicate of FIXED_VALUE_NO_BG
+      key_value_data="--keysize=20 --valuesize=4800"
+      ;;
+  esac
 
   if [ $server_setup -gt 0 ]; then
     ./mutilate --loadonly -s $server_ip $key_value_data
@@ -111,10 +157,11 @@ function run_master {
   ./mutilate $agents -T $mutilate_threads -c $mutilate_connections \
     -s $server_ip -t $mutilate_duration $key_value_data --agent-sampling \
     --save=$RESULT_FILE -w $mutilate_warmup
-}
+  }
 
 # MAIN
 check_env
+echo "Experiment is ${exp_name[$EXPERIMENT]}"
 
 server_setup=1
 if [ "x$1" = "x--no-server-setup" ]
@@ -125,7 +172,7 @@ fi
 if [ $server_setup -gt 0 ]; then
   # echo "How many server cores?"
   # read server_cores
-
+  echo "Setting up server machine..."
   ssh $EXP_SSH_USER@$server_ip "$(set); run_memcached_server" &> /dev/null &
   sleep 1
 fi
@@ -146,6 +193,7 @@ do
       sleep 1
     done
     echo "$m is finished."
+    pkill mutilate
 EOF
 done
